consolidated projects
