+++ DXHRDCHudScaler.Core.Tests/TestUninstallService.cs
@@ -1,6 +1,6 @@
 ﻿using DXHRDCHudScaler.Core.Services;
+using DXHRDCHudScaler.Core.Services.FindDXHRDCExeService;
 using DXHRDCHudScaler.Core.Services.UninstallService;
-using DXHRDCHudScaler.Windows;
 
 namespace DXHRDCHudScaler.Core.Tests;
 
@@ -9,7 +9,7 @@ public class TestUninstallService
     [Test]
     public void TestCanUninstall()
     {
-        var findService = new FindDxhrdcExeService();
+        var findService = new FindDxhrdcExeServiceWin();
         if (findService.TryFind(out var path))
         {
             var sut = new UninstallService();
@@ -20,7 +20,7 @@ public class TestUninstallService
     [Test]
     public void TestUninstall()
     {
-        var findService = new FindDxhrdcExeService();
+        var findService = new FindDxhrdcExeServiceWin();
         if (findService.TryFind(out var path))
         {
             var sut = new UninstallService();
