+++ DXHRDCHudScaler.Core/NativeMethods.txt
