+++ DXHRDCHudScaler.Core/Services/FindDXHRDCExeService/FindDXHRDCExeServiceLinux.cs
@@ -1,0 +1,10 @@
+namespace DXHRDCHudScaler.Core.Services.FindDXHRDCExeService;
+
+public class FindDXHRDCExeServiceLinux : IFindDxhrdcExeService
+{
+    public bool TryFind(out string path)
+    {
+        path = "";
+        return false;
+    }
+}
