--- DXHRDCHudScaler.Windows.Tests/GlobalUsings.cs
@@ -1,1 +1,0 @@
-global using NUnit.Framework;
