+++ DXHRDCHudScaler.sln
@@ -4,13 +4,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DXHRDCHudScaler", "DXHRDCHu
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DXHRDCHudScaler.Core", "DXHRDCHudScaler.Core\DXHRDCHudScaler.Core.csproj", "{A6C71D6E-0754-41BA-B930-D32C4E72CF61}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DXHRDCHudScaler.Windows", "DXHRDCHudScaler.Windows\DXHRDCHudScaler.Windows.csproj", "{9768D4D0-57DB-4D9D-8398-7BE82937EBA7}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DXHRDCHudScaler.Windows.Tests", "DXHRDCHudScaler.Windows.Tests\DXHRDCHudScaler.Windows.Tests.csproj", "{AE023AED-85B8-44C5-83AF-22E4AFCD4458}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DXHRDCHudScaler.Core.Tests", "DXHRDCHudScaler.Core.Tests\DXHRDCHudScaler.Core.Tests.csproj", "{A73C598B-9E4F-4A3B-8A8A-E5E7E8BB5079}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DXHRDXHudScaler.Linux", "DXHRDXHudScaler.Linux\DXHRDXHudScaler.Linux.csproj", "{FBA096EF-697F-486B-A41E-E37FFA686B46}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -26,21 +20,9 @@ Global
 		{A6C71D6E-0754-41BA-B930-D32C4E72CF61}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A6C71D6E-0754-41BA-B930-D32C4E72CF61}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A6C71D6E-0754-41BA-B930-D32C4E72CF61}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9768D4D0-57DB-4D9D-8398-7BE82937EBA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9768D4D0-57DB-4D9D-8398-7BE82937EBA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9768D4D0-57DB-4D9D-8398-7BE82937EBA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9768D4D0-57DB-4D9D-8398-7BE82937EBA7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AE023AED-85B8-44C5-83AF-22E4AFCD4458}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AE023AED-85B8-44C5-83AF-22E4AFCD4458}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AE023AED-85B8-44C5-83AF-22E4AFCD4458}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AE023AED-85B8-44C5-83AF-22E4AFCD4458}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A73C598B-9E4F-4A3B-8A8A-E5E7E8BB5079}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A73C598B-9E4F-4A3B-8A8A-E5E7E8BB5079}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A73C598B-9E4F-4A3B-8A8A-E5E7E8BB5079}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A73C598B-9E4F-4A3B-8A8A-E5E7E8BB5079}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FBA096EF-697F-486B-A41E-E37FFA686B46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FBA096EF-697F-486B-A41E-E37FFA686B46}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FBA096EF-697F-486B-A41E-E37FFA686B46}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FBA096EF-697F-486B-A41E-E37FFA686B46}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
