+++ DXHRDCHudScaler/ViewModels/MainTabViewModel.cs
@@ -8,6 +8,9 @@ using Avalonia.Platform.Storage;
 using DXHRDCHudScaler.Core.Models;
 using DXHRDCHudScaler.Core.Services;
 using DXHRDCHudScaler.Core.Services.BackupService;
+using DXHRDCHudScaler.Core.Services.FindDXHRDCExeService;
+using DXHRDCHudScaler.Core.Services.GetGameRenderResolutionServiceService;
+using DXHRDCHudScaler.Core.Services.ResolutionService;
 using DXHRDCHudScaler.Core.Services.UiScalePatchService;
 using DXHRDCHudScaler.Core.Services.UninstallService;
 using DXHRDCHudScaler.Models;
